chore: organize additional classes into folders
